Moved tests out for CNN and ConvLSTM models
